новый активити с картой ,сделан кривой переход на него из seconactivity (нажимать надо на текст "местоположение")
