+++ app/src/main/AndroidManifest.xml
@@ -40,6 +40,8 @@
         </activity>
         <activity android:name=".SecondActivity">
         </activity>
+        <activity android:name=".MapActivity">
+        </activity>
     </application>
 
 </manifest>
