--- app/src/main/java/googleMap.java
@@ -1,27 +1,0 @@
-import android.app.Activity;
-import android.os.Bundle;
-
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.MapFragment;
-import com.google.android.gms.maps.OnMapReadyCallback;
-
-import examplecom.geomarkers.R;
-
-/**
- * Created by nikita on 04.08.16.
- */
-public class googleMap extends Activity implements OnMapReadyCallback {
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.map_activity);
-        MapFragment mapFragment = (MapFragment) getFragmentManager()
-                .findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-
-    }
-
-    public void onMapReady(GoogleMap googleMap) {
-
-    }
-}
